rename services to viewmodel
