+++ app/src/main/java/com/example/returnpals/viewmodel/AddressRepository.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.services
+package com.example.returnpals.viewmodel
 
 import com.amplifyframework.datastore.generated.model.Address
