+++ app/src/main/java/com/example/returnpals/viewmodel/Repository.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.services
+package com.example.returnpals.viewmodel
 
 import android.util.Log
 import com.amplifyframework.api.graphql.model.ModelMutation
