+++ app/src/main/java/com/example/returnpals/viewmodel/UsersMongoDbDataModel.kt
@@ -1,4 +1,4 @@
-package com.example.returnpals.services
+package com.example.returnpals.viewmodel
